+++ src/components/Card/Card.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import classnames from 'classnames';
 import './Card.css';
 
 const Card = (props) => {
@@ -9,16 +10,20 @@ const Card = (props) => {
     title,
   } = props;
 
+  const cardClass = classnames('card', {
+    'fallback': !link,
+  });
+
   const coverStyle = {
     backgroundImage: `url(${cover})`,
   };
 
   return (
-    <a className="card" href={link} target="_blank">
+    <a className={cardClass} href={link} target="_blank">
       <div className="card-cover" style={coverStyle} />
       <div className="card-details">
-        <h1 className="card-title">{title}</h1>
-        <p>{content}</p>
+        <h3 className="card-title">{title}</h3>
+        <div>{content}</div>
       </div>
     </a>
   );